pyproject.toml
[project]
name = "Sign_pipeline"
version = "0.1.0"
requires-python = "~=3.12.0"
dependencies = [
    "numpy",
    "pyyaml",
    "yt_dlp[default]",
    "silero-vad",
    "faster-whisper",
    "torchcodec",
    "python-dotenv",
    "ultralytics",
    "rtmlib",
    "opencv-python",
    "h5py",
    "lap>=0.5.12",
    "sortedcontainers",
    "pathos",
    "onnxruntime-gpu",
    "nvidia-cublas-cu12",
    "nvidia-cufft-cu12",
    "nvidia-cuda-runtime-cu12",
]